Adjust things for coc.nvim
I should probably try neovim's built-in LSP stuff later, but I'm not
keen on breaking the thing I'm used to right now. Anyway, add some more
key mappings for coc.nvim, particular tab, shift-tab, enter, and []g to
hop around diagnostics.

Disable parameter hints for coc-rust-analyzer because it's pretty
annoying with how it renders it. Also change the file type from rs to
rust so it auto-formats correctly.

Add coc_global_extensions to ensure coc-pairs and coc-rust-analyzer both
get insalled on setup, as well as removing nvim-autopairs because it
conflicts with other mappings. coc-pairs is a worse plugin (it doesn't
handle closing parens very well when paired with completions), but using
it for now to get by.

Remove lazy-loading of coc.nvim. This is so it doesn't result in
annoying errors when it's not loaded and so I don't have to add exists()
calls to every expr mapping that involves coc.nvim. It was also
inconsistent lazy-loading, so don't bother, coc.nvim isn't that
intrusive for the most part.

Remove C-hjkl in insert mode because they were flaky in different
terminals and just not worth the trouble. Repurpose C-k for code
actions.

Add fns.vim just for the COC mappings.